Create Apple.java
Q3. Given a string, return a new string made of n copies of the first 2 chars of the original string where n is the length of the string. The string may be any length. If there are fewer than 2 chars, use whatever is there.
If the input is “Apple” then the output should be “ApApApApAp”