Create rajrajraj.java
Q5. Given a string and an int n, return a string made of n repetitions of the last n characters of the string. You may assume that n is between 0 and the length of the string, inclusive. For example, if the inputs are “Niraj” and 3, then the output should be “rajrajraj.